% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwpls.R
\name{min_grid.lwpls}
\alias{min_grid.lwpls}
\title{Determine the minimal set of model fits for tuning}
\usage{
\method{min_grid}{lwpls}(x, grid, ...)
}
\arguments{
\item{x}{A \code{\link[=lwpls]{lwpls()}} model specification.}

\item{grid}{A tibble with tuning parameter combinations.}

\item{...}{Not currently used.}
}
\value{
A tibble with the minimal tuning grid and a list column with the
submodel values to predict.
}
\description{
\code{min_grid()} is used by the tune package to fit as few models as possible.
LW-PLS predictions for \code{num_comp} components include those for all
smaller numbers of components, so only the largest \code{num_comp} value of
each combination of the other parameters is fitted.
}
\keyword{internal}
