% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localization.R
\name{localization}
\alias{localization}
\title{Localization parameter for LW-PLS}
\usage{
localization(range = c(-9, 5), trans = scales::transform_log2())
}
\arguments{
\item{range}{A two-element vector holding the \emph{defaults} for the smallest
and largest possible values, respectively. If a transformation is
specified, these values should be in the \emph{transformed units}.}

\item{trans}{A \code{trans} object from the \code{scales} package, such as
\code{\link[scales:transform_log10]{scales::transform_log10()}} or \code{\link[scales:transform_reciprocal]{scales::transform_reciprocal()}}. If not
provided, the default is used which matches the units used in \code{range}.
If no transformation, \code{NULL}.}
}
\value{
A \code{quant_param} object (see \code{\link[dials:new_quant_param]{dials::new_quant_param()}}).
}
\description{
The localization parameter controls how fast the similarity weights of
\code{\link[=lwpls]{lwpls()}} decrease with the distance to the query sample. Small values
give very local models; large values give all training samples similar
weights, so that the model tends to a global PLS model.
}
\details{
The default range, \eqn{2^{-9}}{2^-9} to \eqn{2^{5}}{2^5} on a log-2 scale, follows
the candidate values used by Kaneko for the LW-PLS hyperparameter search.
}
\examples{
localization()
dials::value_seq(localization(), 5)
dials::grid_regular(dials::num_comp(c(1, 5)), localization(), levels = 3)
}
