% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwpls.R
\name{lwpls}
\alias{lwpls}
\title{General Interface for Locally-Weighted Partial Least Squares (LWPLS)}
\usage{
lwpls(mode = "unknown", num_comp = NULL, neighbors = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model.
Possible values for this model are "unknown", "regression", or "classification".}

\item{num_comp}{The number of components to retain in the local PLS models.}

\item{neighbors}{The number of neighbors considered at each prediction.}
}
\description{
General Interface for Locally-Weighted Partial Least Squares (LWPLS)
}
\examples{
lwpls() \%>\%
  parsnip::set_args(num_comp = 2, neighbors = 2) \%>\%
  parsnip::set_engine("rnirs") \%>\%
  parsnip::set_mode("regression") \%>\%
  parsnip::translate()

lwpls() \%>\%
  parsnip::set_args(num_comp = 5, neighbors = 3) \%>\%
  parsnip::set_engine("rnirs") \%>\%
  parsnip::set_mode("classification") \%>\%
  parsnip::translate()

}
