% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwpls.R
\name{lwpls}
\alias{lwpls}
\title{General Interface for Locally-Weighted Partial Least Squares (LWPLS)}
\usage{
lwpls(mode = "unknown", num_comp = NULL, neighbors = NULL)
}
\arguments{
\item{mode}{A single character string for the type of model.
Possible values for this model are "unknown", "regression", or "classification".}

\item{num_comp}{The number of components to retain in the local PLS models.}

\item{neighbors}{The number of neighbors considered at each prediction.}
}
\description{
`lwpls()` is a way to generate a _specification_ of a model before fitting and
allows the model to be created using R. The main arguments for the model are:
\itemize{
 \item \code{num_comp}: The number of components to retain in the local PLS models.
 \item \code{neighbors}: The number of neighbors considered at each prediction.
}
}
\details{
The model can be created using the `fit()` function using the
 following _engines_:
\itemize{
\item \pkg{R}:  `"rnirs"`  (the default)
}
}
\section{Engine Details}{


Engines may have pre-set default arguments when executing the
 model fit call. The possible model calls are shown in the Examples section
 below.
}

\examples{
lwpls(num_comp = 2, neighbors = 2) \%>\%
  set_engine("rnirs") \%>\%
  set_mode("regression") \%>\%
  translate()

lwpls(num_comp = 2, neighbors = 1) \%>\%
  set_engine("rnirs") \%>\%
  set_mode("classification") \%>\%
  translate()

lwpls(num_comp = 6) \%>\%
  set_engine("rnirs") \%>\%
  set_mode("regression") \%>\%
  translate()

lwpls() \%>\%
  set_engine("rnirs") \%>\%
  set_mode("classification") \%>\%
  translate()
}
