% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lwpls.R
\name{update.lwpls}
\alias{update.lwpls}
\title{General Interface for Updating Models Parameters}
\usage{
\method{update}{lwpls}(
  object,
  parameters = NULL,
  num_comp = NULL,
  neighbors = NULL,
  fresh = FALSE,
  ...
)
}
\arguments{
\item{object}{lwpls model specification.}

\item{parameters}{A 1-row tibble or named list with \emph{main}
parameters to update. If the individual arguments are used,
these will supersede the values in \code{parameters}. Also, using
engine arguments in this object will result in an error.}

\item{num_comp}{The number of components to retain in the local PLS models.}

\item{neighbors}{The number of neighbors considered at each prediction.}

\item{fresh}{A logical for whether the arguments should be
modified in-place of or replaced wholesale.}

\item{...}{Not used for \code{update()}.}
}
\description{
General Interface for Updating Models Parameters
}
\examples{
model <- lwpls(neighbors =  3)
model
update(model, neighbors = 1)
update(model, neighbors = 1, fresh = TRUE)

}
